Refactoring client modules
Rename `client` module to `codechecker_client`.
