+++ webserver/client/codechecker_client/credential_manager.py
