+++ webserver/client/codechecker_client/product_helper.py
