+++ webserver/client/codechecker_client/thrift_helper.py
